Page crash fix (#154)
* README update and page crash bugfix

* Update Authors.md

* Update README.md

* Update src/pages/index.tsx

Co-authored-by: Spencer Kaiser <SpencerKaiser@users.noreply.github.com>

* Update index.tsx

So I ended up not getting this to work the way I wanted to, but I'll just go through what I tried. Hopefully it just turns out I'm exceedingly stupid (which is honestly more than likely what's happening here).

I tried abstracting the contents of the `promises.push()` function call that contained the calls for getting and setting the request count. I did this thinking that It would update the page if called from inside the `abandonRequest`  and `closeRequest` functions. I wanted to do this to check if there were any more requests in the queue, and if there wasn't, I wanted to update the page to get the new counts, which would in-turn disable the `Get Next` button. However, I quickly realized that updating these values did nothing in the way of causing the page to take in the changes before the interval time. 

Now by my understanding the `setInterval()` function is what causes the page to update after a certain period of time, however I think I have a gap in understanding exactly how it updates the page. From what I am seeing all the interval does is call `setLastUpdateEpoch()` on every interval, but if that's all it takes to update the page then the page would already be updating itself every time, as it is called already in multiple places (`closeRequest`, `getNext`, etc.).

Ultimately it comes down to: How stupid am I? I also apologize for not expressing these concerns earlier than right before the deadline, I am aware that I'm making a goofball out of myself by asking these questions so late. Nevertheless, I am more than happy to fix this as soon as I possibly can once pointed in the right direction.

* Update src/pages/index.tsx

Co-authored-by: Spencer Kaiser <SpencerKaiser@users.noreply.github.com>

Co-authored-by: Spencer Kaiser <SpencerKaiser@users.noreply.github.com>